src/backend/auth_service/requirements_simple.txt
# Simplified requirements for auth_service
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
argon2-cffi==23.1.0
email-validator==2.1.1
fastapi-csrf-protect==1.0.3
slowapi==0.1.9
httpx==0.28.1
python-multipart==0.0.18
prometheus_client==0.21.1
aiofiles==24.1.0
